chore: update "@react-native-community/bob" to the latest version (#70)
* Update "@react-native-community/bob" dependency to the latest version

This commit fixes the bug with not possible to install "@react-navigation/core" using GitHub repo url on Windows. It's a common use case when you work with fork of this original repo and use its url as dependency in your package.json.

Corresponding commit in "@react-native-community/bob"
https://github.com/react-native-community/bob/commit/88c879c83f85bb4f169e78ec180820c1f2bc98a8#diff-e5d6ed39d4bcfdb44ace83a35d62c1a5.
It was released in v0.4.1

* Update yarn.lock file with new "@react-native-community/bob" dependency version